Created LegislativeSessionSetupManager.java
